+++ src/components/notification/AppNotification.ts
@@ -1,0 +1,6 @@
+import {ReactElement} from 'react';
+
+export type AppNotification = {
+  title: string | ReactElement;
+  content?: string | ReactElement;
+};