+++ src/components/result/Result.tsx
@@ -1,45 +1,21 @@
 import React from 'react';
-import {DataRate} from '../../AppConfig';
+import {DataRate, Region} from '../../AppConfig';
 import HelpTooltip from '../help/HelpTooltip';
-import {fmt, withFullStops} from './helpers';
+import {fmt, fmtSeconds, withFullStops} from './helpers';
 
 type ResultGridProps = {
+  region: Region;
   dr: DataRate;
   size: number;
   airtime: number;
   maxPhyPayloadSize: number;
-  maxDwellTime?: number;
 };
-
-export function fmt(n: number, digits: number) {
-  return n.toLocaleString('en-US', {minimumFractionDigits: digits, maximumFractionDigits: digits});
-}
-
-/**
- * Join non-empty fragments into a sentence, adding full stops when combining
- * more than one fragment.
- */
-export function withFullStops(...lines: (string | false | 0 | undefined)[]) {
-  return (
-    lines.reduce((acc: string, line) => {
-      if (!line) {
-        return acc;
-      }
-      return (
-        acc +
-        (acc && acc.slice(-1) !== '.' ? '. ' : acc ? ' ' : '') +
-        line +
-        (acc && line.slice(-1) !== '.' ? '.' : '')
-      );
-    }, '') || undefined
-  );
-}
 
 /**
  * A single result, showing the data rate (SF, bandwidth), time on air and
  * calculated limits.
  */
-export function Result({dr, size, airtime, maxPhyPayloadSize, maxDwellTime = 0}: ResultGridProps) {
+export function Result({region, dr, size, airtime, maxPhyPayloadSize}: ResultGridProps) {
   // Minimum delay in seconds between two packet starts, for maximum 1% duty cycle
   const dcDelay = airtime / 0.01 / 1000;
 
@@ -47,9 +23,12 @@ export function Result({dr, size, airtime, maxPhyPayloadSize, maxDwellTime = 0}:
   const fapMessages = 30000 / airtime;
   const fapDelay = (24 * 3600) / fapMessages;
 
+  const maxDwellTime = region.maxDwellTime || 0;
   const dwellTimeExceeded = maxDwellTime > 0 && airtime > maxDwellTime;
   const sizeExceeded = dr.maxMacPayloadSize && size > maxPhyPayloadSize;
 
+  // The single `&#x200b;` and double `&#x200b;&#x200b;` markers are used to add
+  // whitespace and newlines when copying; see useClipboard.
   return (
     <>
       <HelpTooltip
@@ -57,41 +36,57 @@ export function Result({dr, size, airtime, maxPhyPayloadSize, maxDwellTime = 0}:
           dr.notes,
           dr.maxMacPayloadSize &&
             (sizeExceeded
-              ? `The full payload of ${size} bytes exceeds the maximum packet size of ${maxPhyPayloadSize} bytes`
-              : `Maximum total packet size ${maxPhyPayloadSize} bytes`)
+              ? `The total payload of ${size} bytes exceeds the maximum packet size of ${maxPhyPayloadSize} bytes for ${region.label} SF${dr.sf}BW${dr.bw}`
+              : `For ${region.label} SF${dr.sf}BW${dr.bw} the maximum total packet size is ${maxPhyPayloadSize} bytes`)
         )}
       >
         <div
           className={`Result-datarate ${
             sizeExceeded ? 'Result-has-warning' : dr.notes ? 'Result-has-note' : ''
           }`}
+          role="cell"
         >
           <div className="Result-dr">{dr.name}</div>
           <div>
+            &#x200b;&#x200b;
             <span className="Result-sf">SF{dr.sf}</span>
+            &#x200b;&#x200b;
             <span className="Result-unit Result-unit-bw">
               BW
               <br />
               {dr.bw}
             </span>
           </div>
-          {sizeExceeded && <div className={'Result-datarate-warning'}>max size exceeded</div>}
+          {sizeExceeded && (
+            <div className={'Result-datarate-warning'}>&#x200b;&#x200b;max size exceeded</div>
+          )}
         </div>
       </HelpTooltip>
 
       <HelpTooltip
-        content={withFullStops(
-          `${fmt(airtime, 3)} milliseconds time on air`,
-          dwellTimeExceeded && `This exceeds the maximum dwell time of ${maxDwellTime} milliseconds`
-        )}
+        content={
+          <>
+            On SF{dr.sf}BW{dr.bw}, a total packet size of {size} bytes{' '}
+            <a href={document.location.href}>needs {fmt(airtime, 2)} milliseconds time on air</a>.
+            {dwellTimeExceeded && (
+              <>
+                {' '}
+                This exceeds the maximum dwell time of {maxDwellTime} milliseconds for{' '}
+                {region.label}.
+              </>
+            )}
+          </>
+        }
       >
-        <div className={`Result-airtime ${dwellTimeExceeded ? 'Result-has-warning' : ''}`}>
+        <div
+          className={`Result-airtime ${dwellTimeExceeded ? 'Result-has-warning' : ''}`}
+          role="cell"
+        >
           <div>
-            {fmt(airtime, 1)}
-            <span className="Result-unit">ms</span>
+            {fmt(airtime, 1)}&#x200b;<span className="Result-unit">ms</span>
           </div>
           {dwellTimeExceeded && (
-            <div className={'Result-airtime-warning'}>max dwell time exceeded</div>
+            <div className={'Result-airtime-warning'}>&#x200b;&#x200b;max dwell time exceeded</div>
           )}
         </div>
       </HelpTooltip>
@@ -100,18 +95,21 @@ export function Result({dr, size, airtime, maxPhyPayloadSize, maxDwellTime = 0}:
         content={
           <>
             If regional legal or LoRaWAN maximum duty cycles apply, then a 1% maximum duty cycle
-            needs a <em>minimum</em> of {fmt(dcDelay, 3)} seconds between <em>any</em> subsequent
+            needs a <em>minimum</em> of {fmt(dcDelay, 2)} seconds between <em>any</em> subsequent
             packet starts in the same subband.
           </>
         }
       >
-        <div className="Result-dutycycle">
+        <div className="Result-dutycycle" role="cell">
           <div>
             {fmt(dcDelay, 1)}
+            &#x200b;
             <span className="Result-unit">sec</span>
+            &#x200b;&#x200b;
           </div>
           <div>
             {fmt(Math.floor(3600 / dcDelay), 0)}
+            &#x200b;
             <span className="Result-unit">
               msg
               <br />
@@ -124,38 +122,52 @@ export function Result({dr, size, airtime, maxPhyPayloadSize, maxDwellTime = 0}:
       <HelpTooltip
         content={
           <>
+            {/*{ Links are only added for copy/pasting rich text. On touch devices links in tooltips
+             would be clickable, but still are rendered invisibly, to not confuse the user clicking
+             the link that points back to this very location. }*/}
             <p>
-              The TTN Fair Access Policy allows for at most 30 seconds time on air per device, per
-              24 hours. An airtime of {fmt(airtime, 3)}&nbsp;ms thus yields a maximum of{' '}
-              {fmt(fapMessages, 1)} messages.
+              The <a href="https://www.thethingsnetwork.org/forum/t/1300">TTN Fair Access Policy</a>{' '}
+              allows for at most 30 seconds time on air per device, per 24 hours. So, an{' '}
+              <a href={document.location.href}>
+                airtime of {fmt(airtime, 1)} milliseconds for {size} bytes on SF{dr.sf}BW{dr.bw}
+              </a>{' '}
+              imposes a limit of {fmt(fapMessages, 1)} messages per day.
             </p>
+            {/* Add newlines between paragraphs in case we're copy/pasting plain text. */}
+            {'\n\n'}
             <p>
-              When transmitting all day, this translates to, <em>on average</em>, a minimum of{' '}
-              {fmt(fapDelay, 1)} seconds between the starts of two uplinks, or a maximum of{' '}
-              {fmt(fapMessages / 24, 1)}&nbsp;messages per hour.
+              When transmitting all day, <strong>on average</strong> this needs a minimum of{' '}
+              {fmtSeconds(fapDelay)} between the starts of two uplinks, or a maximum of{' '}
+              {fmt(fapMessages / 24, 1)}
+              &nbsp;messages per hour.
             </p>
           </>
         }
       >
-        <div className="Result-fairaccess">
+        <div className="Result-fairaccess" role="cell">
           <div className="Result-fairaccess-messages">
             {fmt((24 * 3600) / fapMessages, 1)}
+            &#8203;
             <span className="Result-unit">
-              sec
+              sec&#8203;
               <br />
               (avg)
             </span>
+            &#8203;&#8203;
           </div>
           <div>
             <span className="Result-fairaccess-messages-per-hour">{fmt(fapMessages / 24, 1)}</span>
+            &#8203;
             <span className="Result-unit Result-unit-hour">
               avg
               <br />
               /hour
             </span>
+            &#8203;&#8203;
           </div>
           <div className="Result-fairaccess-messages-per-day">
             {fmt(Math.floor(fapMessages), 0)}
+            &#8203;
             <span className="Result-unit">
               msg
               <br />